fix : installation of tweetnacl-util depedency and correction of getValidated3pid
