+++ package.json
@@ -38,8 +38,9 @@
     "test": "lerna run test"
   },
   "dependencies": {
+    "@remix-run/express": "^1.16.0",
     "express": "^4.18.2",
-    "@remix-run/express": "^1.16.0"
+    "tweetnacl-util": "^0.15.1"
   },
   "devDependencies": {
     "@crowdsec/express-bouncer": "^0.1.0",
@@ -72,7 +73,7 @@
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "js-nacl": "^1.4.0",
-    "lerna": "^7.1.4",
+    "lerna": "^8.1.3",
     "moment": "^2.29.4",
     "node-fetch": "^3.3.0",
     "node-fetch-jest": "npm:node-fetch@^2.6.9",
