+++ packages/matrix-identity-server/src/index.test.ts
@@ -438,21 +438,31 @@ describe('Use configuration file', () => {
     describe('/_matrix/identity/v2/3pid', () => {
       describe('/_matrix/identity/v2/3pid/getValidated3pid', () => {
         let sid: string, token: string
+        it('should reject missing parameters', async () => {
+          const response = await request(app)
+            .get('/_matrix/identity/v2/3pid/getValidated3pid')
+            .query({
+              client_secret: 'mysecret'
+            })
+            .set('Authorization', `Bearer ${validToken}`)
+            .set('Accept', 'application/json')
+          expect(response.statusCode).toBe(400)
+          expect(response.body.errcode).toBe('M_MISSING_PARAMS')
+        })
         it('should return 404 if no valid session is found', async () => {
           const response = await request(app)
             .get('/_matrix/identity/v2/3pid/getValidated3pid')
+            .query({
+              client_secret: 'invalidsecret',
+              sid: 'invalidsid'
+            })
             .set('Authorization', `Bearer ${validToken}`)
             .set('Accept', 'application/json')
-            .send({
-              client_secret: 'invalid_secret',
-              sid: 'invalid_sid'
-            })
           expect(response.body.errcode).toBe('M_NO_VALID_SESSION')
           expect(response.statusCode).toBe(404)
         })
-        // Necessary test to get the sid and token
-        test('copy of requestToken test', async () => {
-          const response = await request(app)
+        it('should return 400 if the session is not validated', async () => {
+          const responseRequestToken = await request(app)
             .post('/_matrix/identity/v2/validate/email/requestToken')
             .set('Authorization', `Bearer ${validToken}`)
             .set('Accept', 'application/json')
@@ -462,29 +472,28 @@ describe('Use configuration file', () => {
               next_link: 'http://localhost:8090',
               send_attempt: 1
             })
-          expect(response.statusCode).toBe(200)
+          expect(responseRequestToken.statusCode).toBe(200)
           expect(sendMailMock.mock.calls[0][0].to).toBe('xg@xnr.fr')
           expect(sendMailMock.mock.calls[0][0].raw).toMatch(
             /token=([a-zA-Z0-9]{64})&client_secret=newsecret&sid=([a-zA-Z0-9]{64})/
           )
           token = RegExp.$1
           sid = RegExp.$2
-        })
-        it('should return 400 if the session is not validated', async () => {
+
           const response = await request(app)
             .get('/_matrix/identity/v2/3pid/getValidated3pid')
             .set('Authorization', `Bearer ${validToken}`)
-            .set('Accept', 'application/json')
-            .send({
+            .query({
               client_secret: 'newsecret',
               sid
             })
+            .set('Accept', 'application/json')
           expect(response.body.errcode).toBe('M_SESSION_NOT_VALIDATED')
           expect(response.statusCode).toBe(400)
         })
-        // Necessary test to validate the session
-        test('copy of submitToken test', async () => {
-          const response = await request(app)
+        /* Works if the validationTime is set to 0 millisecond in 3pid/getValidated3pid.ts 
+        it('should return 400 if the session is expired', async () => {
+          const responseSubmitToken = await request(app)
             .get('/_matrix/identity/v2/validate/email/submitToken')
             .query({
               token,
@@ -492,35 +501,43 @@ describe('Use configuration file', () => {
               sid
             })
             .set('Accept', 'application/json')
-          expect(response.body).toEqual({ success: true })
-          expect(response.statusCode).toBe(200)
-        })
-        /* Works if the validationTime is set to 1 millisecond in 3pid/index.ts
-        it('should return 400 if the session is expired', async () => {
+          expect(responseSubmitToken.body).toEqual({ success: true })
+          expect(responseSubmitToken.statusCode).toBe(200)
           const response = await request(app)
             .get('/_matrix/identity/v2/3pid/getValidated3pid')
             .set('Authorization', `Bearer ${validToken}`)
-            .set('Accept', 'application/json')
-            .send({
+            .query({
               client_secret: 'newsecret',
               sid
             })
+            .set('Accept', 'application/json')
           expect(response.body.errcode).toBe('M_SESSION_EXPIRED')
           expect(response.statusCode).toBe(400)
         })
         */
         it('should return 200 if a valid session is found', async () => {
-          const response = await request(app)
-            .get('/_matrix/identity/v2/3pid/getValidated3pid')
-            .set('Authorization', `Bearer ${validToken}`)
-            .set('Accept', 'application/json')
-            .send({
+          const responseSubmitToken = await request(app)
+            .get('/_matrix/identity/v2/validate/email/submitToken')
+            .query({
+              token,
               client_secret: 'newsecret',
               sid
             })
+            .set('Accept', 'application/json')
+          expect(responseSubmitToken.body).toEqual({ success: true })
+          expect(responseSubmitToken.statusCode).toBe(200)
+          const response = await request(app)
+            .get('/_matrix/identity/v2/3pid/getValidated3pid')
+            .set('Authorization', `Bearer ${validToken}`)
+            .query({
+              client_secret: 'newsecret',
+              sid
+            })
+            .set('Accept', 'application/json')
           expect(response.statusCode).toBe(200)
         })
       })
+
       describe('/_matrix/identity/v2/3pid/bind', () => {
         it('should find the 3pid - matrixID association after binding', async () => {
           const response_request_token = await request(app)
