+++ packages/matrix-identity-server/src/index.ts
@@ -38,7 +38,7 @@ import UserDB from './userdb'
 import _validateMatrixToken from './utils/validateMatrixToken'
 import RequestToken from './validate/email/requestToken'
 import SubmitToken from './validate/email/submitToken'
-import GetValidated3pid from './3pid'
+import GetValidated3pid from './3pid/getValidated3pid'
 import unbind from './3pid/unbind'
 import bind from './3pid/bind'
 export { type tokenContent } from './account/register'
